Extracted contributor list into separate component
